+++ CHANGELOG.md
@@ -5,8 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Feature GeoDistance is implemented.
+- Add Geometry type support.
 
 ### Fixed
+- Fix geography-functions issues.
 
 ### Changed